.history/requirement_20260317091505.txt
opencv-python
reportlab
requests
pandasopenpyxl

psql -h localhost -U ams_admin -d ams_db || sudo -u postgres psql